@dim 2016-03-11
```
$ benchcmp old.txt new.txt 
benchmark                         old ns/op     new ns/op     delta
BenchmarkParseReplyStatus-4       123           113           -8.13%
BenchmarkParseReplyInt-4          102           102           +0.00%
BenchmarkParseReplyError-4        227           215           -5.29%
BenchmarkParseReplyString-4       170           178           +4.71%
BenchmarkParseReplySlice-4        901           889           -1.33%
BenchmarkAppendArgs-4             227           218           -3.96%
BenchmarkRedisClusterPing-4       9100          9183          +0.91%
BenchmarkPool-4                   241           308           +27.80%
BenchmarkRedisPing-4              7638          7528          -1.44%
BenchmarkRedisSet-4               18755         18934         +0.95%
BenchmarkRedisGetNil-4            7676          7668          -0.10%
BenchmarkRedisSetGet64Bytes-4     16940         16321         -3.65%
BenchmarkRedisSetGet1KB-4         17856         17755         -0.57%
BenchmarkRedisSetGet10KB-4        40290         37613         -6.64%
BenchmarkRedisSetGet1MB-4         2195778       2157348       -1.75%
BenchmarkRedisSetGetBytes-4       43140         37612         -12.81%
BenchmarkRedisMGet-4              8281          7888          -4.75%
BenchmarkSetExpire-4              19366         16504         -14.78%
BenchmarkPipeline-4               9863          9942          +0.80%
BenchmarkZAdd-4                   8951          8828          -1.37%

benchmark                         old allocs     new allocs     delta
BenchmarkParseReplyStatus-4       1              1              +0.00%
BenchmarkParseReplyInt-4          1              1              +0.00%
BenchmarkParseReplyError-4        2              2              +0.00%
BenchmarkParseReplyString-4       1              1              +0.00%
BenchmarkParseReplySlice-4        8              8              +0.00%
BenchmarkAppendArgs-4             0              0              +0.00%
BenchmarkRedisClusterPing-4       4              4              +0.00%
BenchmarkPool-4                   0              0              +0.00%
BenchmarkRedisPing-4              4              4              +0.00%
BenchmarkRedisSet-4               7              7              +0.00%
BenchmarkRedisGetNil-4            5              5              +0.00%
BenchmarkRedisSetGet64Bytes-4     11             11             +0.00%
BenchmarkRedisSetGet1KB-4         11             11             +0.00%
BenchmarkRedisSetGet10KB-4        12             12             +0.00%
BenchmarkRedisSetGet1MB-4         13             13             +0.00%
BenchmarkRedisSetGetBytes-4       12             12             +0.00%
BenchmarkRedisMGet-4              13             13             +0.00%
BenchmarkSetExpire-4              13             13             +0.00%
BenchmarkPipeline-4               17             17             +0.00%
BenchmarkZAdd-4                   8              8              +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkParseReplyStatus-4       32            32            +0.00%
BenchmarkParseReplyInt-4          8             8             +0.00%
BenchmarkParseReplyError-4        32            32            +0.00%
BenchmarkParseReplyString-4       32            32            +0.00%
BenchmarkParseReplySlice-4        170           170           +0.00%
BenchmarkAppendArgs-4             0             0             +0.00%
BenchmarkRedisClusterPing-4       116           116           +0.00%
BenchmarkPool-4                   0             0             +0.00%
BenchmarkRedisPing-4              116           116           +0.00%
BenchmarkRedisSet-4               10707         10707         +0.00%
BenchmarkRedisGetNil-4            160           160           +0.00%
BenchmarkRedisSetGet64Bytes-4     418           418           +0.00%
BenchmarkRedisSetGet1KB-4         1378          1378          +0.00%
BenchmarkRedisSetGet10KB-4        21349         21349         +0.00%
BenchmarkRedisSetGet1MB-4         2113967       2113942       -0.00%
BenchmarkRedisSetGetBytes-4       21365         21365         +0.00%
BenchmarkRedisMGet-4              352           352           +0.00%
BenchmarkSetExpire-4              380           380           +0.00%
BenchmarkPipeline-4               800           800           +0.00%
BenchmarkZAdd-4                   224           224           +0.00%
```