@vmihailenco 2016-03-12
>we end up with more and more pool timeouts

Can you share DiatTimeout and ReadTimeout? I suspect that you are affected by combination of:
- read timeout -> connection is closed
- dial timeout -> new connection can't be opened

>I suspect it's due to the iterations on connList.Remove and connList.Replace

See https://github.com/go-redis/redis/pull/273. Remove is not very fast, but it is fast enough and does not become slower when pool size grows.

>As these happen within a mutex, a failing connection effectively takes down everything else, causing the pool to block.

We hold lock only for adding/removing connection from the list. Lock is not held when connection is recycled.

>It also means that we round-robin across all our connections in the pool and large pools never actually shrink.

True, but this does not look very important for Redis which can handle thousands of connections... I would like to fix this, but honestly I don't see how it can fix your issue.