@dim 2016-03-12
Sure, I can give you some background. The settings are pretty standard (and generous):
```
opt := redis.ClusterOptions{
  PoolSize: 20,
  DialTimeout: 5*time.Second,
  ReadTimeout: time.Second,
  WriteTimeout: time.Second,
  PoolTimeout: time.Second,
}
```

Every now and then, we see "spikes" of `redis: connection pool timeout`, although there is no indication of a break-down on operational level. The cluster nodes are up, RDB/AOF are disabled, and there is nothing in the redis logs or in SLOWLOG that would indicate a "lock up". There is nothing that would suggest a network problem either.