@dim 2016-03-12
The odd thing we have identified is that the frequency of these spikes is increasing when we increase the `PoolSize`. We bumped the value from `20` to `200` and the pool time outs instantly went through the roof.