@dim 2016-03-12
> True, but this does not look very important for Redis which can handle thousands of connections...

I don't think it's important to Redis, but it's important to the Go client. Assuming a PoolSize of 10 connections with a cluster of 200 nodes, you (worst case) end up with 2,000 open TCP connections and none of these will ever be retired.