@vmihailenco 2016-03-12
Added your changes in https://github.com/go-redis/redis/pull/277