@vmihailenco 2016-03-12
Tagged v3.5.0 with fixes from this PR.